requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
langchain
python-dotenv
langchain-community